tests/wpt/web-platform-tests/largest-contentful-paint/loadTime-after-appendChild.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: delayed appended image.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(function (t) {
    if (!window.LargestContentfulPaint) {
      assert_unreached("LargestContentfulPaint is not implemented");
    }
    let beforeLoad;
    const observer = new PerformanceObserver(
      t.step_func_done(entryList => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        assert_equals(entry.entryType, 'largest-contentful-paint');
        assert_equals(entry.startTime, entry.renderTime, 'startTime should equal renderTime');
        assert_equals(entry.duration, 0);
        assert_equals(entry.url, window.location.origin + '/images/black-rectangle.png');
        assert_greater_than(entry.renderTime, entry.loadTime,
          'The image render time should occur after it is appended to the div.');
        assert_greater_than(entry.loadTime, beforeLoad,
          'The image load timestamp should occur after script starts running.');
        assert_less_than(entry.renderTime, performance.now(),
          'Image render time should be before the observer callback is executed.')
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
    const img = document.createElement('img');
    img.src = '/images/black-rectangle.png';
    t.step_timeout(() => {
      beforeLoad = performance.now();
      document.getElementById('image_div').appendChild(img);
    }, 200)
  }, 'Image loadTime occurs after appendChild is called.');
</script>
<div id='image_div'></div>
</body>
